+++ libs/gbd-mapping/CHANGELOG.rst
@@ -1,3 +1,12 @@
+**7.0.0 - 07/27/26**
+
+- Remove ``gbd_mapping_generator`` submodule and related tests and documentation
+
+**6.0.7 - 07/23/26**
+
+- Streamline package __init__.py docstrings
+- Add project.urls to pyproject.toml
+
 **6.0.6 - 07/14/26**
 
 - Update stale references for monorepo libraries